Move class IoUtils into utils package
